Move test folder near components
